Final README
In March 2013, Kohli started a charity foundation called Virat Kohli Foundation (VKF). The organisation aims at helping underprivileged kids and conducts events to raise funds for the charity. According to Kohli, the foundation works with select NGOs to "create awareness, seek support and raise funds for the various causes they endorse and the philanthropic work they engage in."In May 2014, eBay and Save the Children India conducted a charity auction with VKF, with its proceeds benefiting the education and healthcare of underprivileged children.